pyproject.toml
[project]
name = "step-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dtaidistance>=2.3.13",
    "fastplotlib[imgui,notebook]>=0.5.1",
    "imgui-bundle>=1.6.0,<1.92.0",
    "ipykernel>=6.29.5",
    "ipympl>=0.9.7",
    "jupyter-rfb>=0.5.1",
    "matplotlib>=3.10.0",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "pyarrow>=18.1.0",
    "pyqt6>=6.9.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "seaborn>=0.13.2",
    "stumpy>=1.13.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipympl>=0.9.6",
    "setuptools",
]
